This is the first part of a new feature: You can get the file contents via getValue() method. It avoids the use of the file field in the standard way, sorry.
The second part is to build the edit part for the field type, but a proposal is included in the source code.

I have wasted so many spare time to try to figure out how to accomplish, but I am a sysadmin&php developer and my javascript is not very strong when multiple inheritance is in play.

I will be happy to continue contributing when my spare time lets it.